fix: address review findings across voucher normalization contracts
- Replace panic-on-overflow .unwrap() with error propagation in factory
  LP token minting/burning (ack_and_timeout.rs L190,206,401,495)
- Gracefully skip dust amounts in _release_voucher instead of aborting
  entire multi-recipient transaction
- Add Phase 2 idempotency guard and strict single-metadata assertion
  in virtual_balance migration
- Remove dead ensure! in pool creation Ok(metadata) arm
- Remove dead .try_into() on already-Uint256 balance amount
- Use voucher amount directly for voucher partner fees (skip unnecessary
  metadata query that could crash on unregistered chain)
- Switch VLP migrations from .load() to .may_load() for clarity
- Fix misleading ZeroAssetAmount error on Dynamic limit validation

+++ packages/euclid/src/normalize.rs
@@ -24,10 +24,10 @@ pub fn normalize(
         return Ok(amount);
     }
     if to_decimals > from_decimals {
-        let factor = Uint256::from(10u128).checked_pow((to_decimals - from_decimals) as u32)?;
+        let factor = Uint256::from(10u128).checked_pow(to_decimals - from_decimals)?;
         Ok(amount.checked_mul(factor)?)
     } else {
-        let factor = Uint256::from(10u128).checked_pow((from_decimals - to_decimals) as u32)?;
+        let factor = Uint256::from(10u128).checked_pow(from_decimals - to_decimals)?;
         Ok(amount.checked_div(factor)?)
     }
 }
